add methods between article and palet

+++ src/main/java/pl/pjatk/store/service/ArticleService.java
@@ -21,30 +21,39 @@ public class ArticleService {
     public Optional<Article> findById(int id) throws ArticleException {
         if (articleRepository.findById(id).isPresent()) {
             return articleRepository.findById(id);
-        } else throw new ArticleException(id);
+        } else throw new ArticleException();
     }
 
     public List<Article> findAll() {
         return articleRepository.findAll();
     }
 
-    public Article addNew(Article player) {
-        return articleRepository.save(player);
+    public Article addNew(Article article) {
+        return articleRepository.save(article);
     }
 
-    public void deleteByID(int id) {
-        articleRepository.deleteById(id);
+    public void deleteByID(int id) throws ArticleException {
+        if (articleRepository.findById(id).isPresent()) {
+            articleRepository.deleteById(id);
+            if (articleRepository.findById(id).isPresent()) {
+                throw new ArticleException("Nie usunął się");
+            }
+        } else throw new ArticleException("Artykuł nie istnieje");
     }
 
-    public Optional<Article> updateById(int id, String name, int price, int weight) {
+    public Optional<Article> updateById(int id, String name, int price, int weight) throws ArticleException {
         Optional<Article> updateArticle = articleRepository.findById(id);
         updateArticle.get().setName(name);
         updateArticle.get().setPrice(price);
         updateArticle.get().setWeight(weight);
-        return updateArticle;
+        if (updateArticle.get().getName().equals(name) &&
+                updateArticle.get().getPrice() == price &&
+                updateArticle.get().getWeight() == weight)
+            return updateArticle;
+        else throw new ArticleException("Nie zaktualizowano artykułu");
     }
 
-    public List<Article> updateByName(String uniqueName, String name, int price, int weight) {
+    public List<Article> updateByName(String uniqueName, String name, int price, int weight) throws ArticleException {
         List<Article> listOfAll = articleRepository.findAll();
         List<Article> updateArticles = new ArrayList<>();
         for (Article article : listOfAll) {
@@ -55,6 +64,8 @@ public class ArticleService {
                 updateArticles.add(article);
             }
         }
+        if (updateArticles.size()==0)
+            throw new ArticleException("Nie znaleziono artykułów");
         return updateArticles;
     }
 }
